extract SelectPropTypes into its own file and default export
Due to how the ES6 code is being transpiled, the exports statement in the Select component lib file were breaking tests.
