+++ src/index.js
@@ -2,6 +2,7 @@ import Select from './Select';
 import Async from './Async';
 import AsyncCreatable from './AsyncCreatable';
 import Creatable from './Creatable';
+import SelectPropTypes from './propTypes/SelectPropTypes';
 
 Select.Async = Async;
 Select.AsyncCreatable = AsyncCreatable;
@@ -11,5 +12,6 @@ export default Select;
 export {
 	Async,
 	AsyncCreatable,
-	Creatable
+	Creatable,
+	SelectPropTypes
 };
